mouse stationary parameter added. moved hud related functions into hud.rs
