+++ src/camera/mod.rs
@@ -1,22 +1,19 @@
-use std::vec;
-mod hud;
+pub mod hud;
 
 use crate::{
     engine::vector2::Vector2,
-    gpustoredinstances::GpuStoredHUDObject,
     input::Input,
 };
 use hud::HudObject;
-
-use self::hud::HudFlag;
 
 pub struct Camera {
     pub coordinates: Vector2, //The cameras coordinates are the coordinates of the tile in the center
     pub target_coordinates: Vector2,
     pub tiles_fit: Vector2,
     pub target_tiles_fit: Vector2,
-    pub camera_size: Vector2,
-    hud_objects: Vec<HudObject>,
+    pub camera_size: Vector2,   //Number represents pixels on both axis.
+    pub hud_objects: Vec<HudObject>,
+    //All hud related things are found in hud.rs, not here.
 }
 
 impl Camera {
@@ -69,7 +66,6 @@ impl Camera {
     pub fn window_resized(&mut self, new_screen_size: Vector2) {
         self.target_tiles_fit = self.target_tiles_fit / (self.camera_size / new_screen_size);
         self.tiles_fit = self.target_tiles_fit;
-        self.camera_size = new_screen_size;
     }
 
     pub fn ease_to_tile(&mut self, coordinates: Vector2) {
@@ -87,37 +83,6 @@ impl Camera {
         let y = -screen_position.x / (2.0 / self.tiles_fit.x)
             + screen_position.y * 2.0 / (2.0 / self.tiles_fit.y);
         Vector2::new(x, y) + self.coordinates
-    }
-
-    pub fn get_hud_object_at_screen_position(
-        &self,
-        screen_position: Vector2,
-    ) -> Option<&HudObject> {
-        for hud_object in &self.hud_objects {
-            if hud_object.screen_position_inside_hud(screen_position) {
-                return Some(hud_object);
-            }
-        }
-        None
-    }
-
-    pub fn get_hud_instance_coordinates(&self) -> Vec<GpuStoredHUDObject> {
-        let mut gpu_stored_hud_objects =
-            vec::from_elem(GpuStoredHUDObject::zero(), self.hud_objects.len());
-        for (vector_index, hud_object) in self.hud_objects.iter().enumerate() {
-            if hud_object.is_shown() {
-                gpu_stored_hud_objects[vector_index] = GpuStoredHUDObject {
-                    screen_position: [
-                        hud_object.top_left.x,
-                        hud_object.top_left.y,
-                        hud_object.z_layer as f32,
-                    ],
-                    object_size: (hud_object.bottom_right - hud_object.top_left).into(),
-                    texture_layer: 0,
-                }
-            }
-        }
-        gpu_stored_hud_objects
     }
 
     //pub fn tile_coordinates_to_screen_position(&self, coordinates: Vector2) -> Vector2 {
